Add build framework for python extension
Implement the minimum required framework to build a pybind11 [1]
generated python extension. The approach is based on various Equinor
projects [2].

The python building relies on high-quality export files, both from
install(EXPORT) and export(TARGETS) for both installed and build-dir
copies of openvds.

In short, the approach is to build the python code with python tooling,
integrated with the root cmake project through a custom target.

[1] https://pybind11.readthedocs.io/en/stable/
[2] https://github.com/equinor/ecl3