Add roundtrip integration test
if cmake parameters TEST_SEGY_FILE and TEST_URL with optional
TEST_CONNECTION then the test will used said parameters import the SEGY
file with default parameters, and then export the VDS to a local SEGY
file. It will then do a SHA256 hash of both files and validate that the
hashes er the same.